move test_scenario up, delete scenario module
