+++ src/lib.rs
@@ -1,7 +1,7 @@
 pub mod db;
 pub mod error;
 pub mod generator;
-pub mod scenario;
 pub mod spammer;
+pub mod test_scenario;
 
 pub type Result<T> = std::result::Result<T, error::ContenderError>;
