--- src/scenario/mod.rs
@@ -1,1 +1,0 @@
-pub mod test_scenario;
